Include Schema in runtime resource
The schema is necessary to get the top level fields of a resource,
which is necessary when running readyOn expressions. In the future
we can store the top level fields and remove the schema if needed.

Also adding a new resource variable kind for readyOn expressions,
to prevent the resolveDynamicVariables function from evaluating them.