(feat) Add conditional expressions for resources
Conditional expressions allow the user to dictate whether they want to create a resource or not via a series of expressions that evaluate to a boolean. The variables in the expressions need to be declared in the instance spec for now.
If all the expressions evaluate to true, we will create the resource, if at least one returns false, we will ignore the resource and all its dependents.