Add support for (custom) types
This adds the `Types` field to the ResourceGraphDefinition CRD. It
contains a map of names to type definitions, where the definitions
follow the simple schema.

The defined types are then loaded into the transformer as predefined
types. Appearances in the spec are then properly resolved.

An example for usage would be:

``` yaml
apiVersion: kro.run/v1alpha1
kind: ResourceGraphDefinition
metadata:
  name: service-entry-dns
spec:
  schema:
    apiVersion: v1alpha1
    kind: ServiceEntryDNS
    spec:
      ports: "[]port | required=true"
    types:
      port:
         name: string
         protocol: string
         number: integer | required
```

Known issues: Currently there is no support for references of custom
types within the definition of a custom type!

Fixes #144