Fix (custom) simple types with required marker
If a simple type was used with the required marker:

``` yaml
types:
  myType: string | required=true description="my description"
spec:
  myValue: myType
```
the required marker was lost, while the description was set properly.
This happened as in the applyMarkers method of the transformer the
required information is set to the parent.

While loading the predefined types this parent is just a shell object,
that is later read to fill the list of predefined types and then
discarded.

Now the information on required types is captured in the
`predefinedType` type and then later used to set the parent object
required fields when handling predefined types.