chore: wrap RGD compile errors
These changes introduce error wrapping during RGD building.
Currently if there is any error in RGD, it is returned immidiately.
When we introduce a CLI, we want to be able to see all the errors in the
RGD at once, instead of needing to run the `validate` command multiple
times for every single error in the RGD.