feat: Add support for custom CEL functions
This adds support for a new top-level field on the RGD that enables
users to create their own functions that they can use in CEL
expressions. It currently only supports primitive types like strings and
ints, but can be extended for more complex types in the future. Adding
support for other runtimes is possible, but is left for a future
addition. One idea I have in mind is to refer to an external resource
that can be used to execute the functions, but one could imagine
embedding starlark and wasm runtimes in kro itself and run functions
embedded verbatim in the RGD.

breaking module API change: I removed the Inspector constructor that did
not take an environment as it wasn't used other than outside of tests
and was signaling that an environment wasn't necessary when it created
one on its own anyway.

I also updated the Inspector to look at all functions declared in the
environment, not only those in the static allow list. This resulted in
some tests needing to be updated as they were referring to functions
added to the environment and not in the allow list (such as duration).

For environment, I updated it to store resourceIDs directly as custom
declarations, to avoid having to duplicate logic in the new
krocel.Extend function that extends a CEL environment instead of
creating a new one. Extending the environment reduces some duplication
in builder.go and I imagine a similar pattern can be used in the future
with strongly typed resources being added to the environment
(effectively having an environment builder that has an understanding of
the schema and resource types).

As for the functions themselves, they are topologically sorted so that
one function can refer to another, not caring about which order the
functions are declared. Functions support overloads by input and output
parameters.

Closes #361.