fix: make sure that dynamic controller informer queues, workers are shutdown correctly
previously the informer queue of the dynamic controller and the manager shut-down immediately (killed). This is problematic because it means that a reconcile loop can potentially be interrupted in the middle of a run even though we have a normally graceful shutdown.

This fixes this by adding the dynamic controller into the manager runnables instead of leaking a goroutine as well as fixing how the manager is actually stopped (previously the manager was run in a goroutine).

Additionally we now properly shutdown the queue of the dynamic controller and wait for the workers to finish processing.

Note that if this process still takes longer than the graceful shutdown period, we forcefully kill the manager and the binary.

Because the shutdown timeout is now respected globally, I have replaced `--dynamic-controller-default-shutdown-timeout` and `KRO_DYNAMIC_CONTROLLER_DEFAULT_SHUTDOWN_TIMEOUT` with `--graceful-shutdown-timeout` and `KRO_GRACEFUL_SHUTDOWN_TIMEOUT` because the shutdown timeout is now handled globally within the manager and not individually just for the dynamic controller.

Signed-off-by: Jakob Möller <jakob.moeller@sap.com>