Publish static kro manifests on release
The industry tends to have different descriptions for
helm and plain k8s manifests with kustomize.

We can notice it in argoCD where they maintain the
[manifests](https://github.com/argoproj/argo-cd/blob/master/manifests/core-install.yaml)
as well as the helm [chart](https://github.com/argoproj/argo-helm/tree/main/charts/argo-cd)
(even in a different repository).

When we first tried to generate static manifests using this technique,
during the review, a concern was raised about the additional cognitive
load and maintenance burden having 2 different deployment methods would introduce.

This commit aims at providing static bundles to benefit from static manifests without
the additional maintenance burden.
It keeps Helm as the source of truth for deployment description and picks a list of well-known
values combination to generate static manifests out of it.