fix: remove RGD owner ID check to allow RGDs to adopt their CRD
Currently, when a user deletes an RGD, and recreates it, the new RGD
is no longer able to adopt the CRD the first RGD left dangling.

This change removes the RGD ID check, ensuring, if an RGD with the same
name as the one that created the CRD is created, it can adopt it.