fix: prevent panic due to missing runtime
Includes a new integration test to validate that an invalid resource surfaces a failure condition with the expected "Inactive" state. Refactors status updates to no longer rely on the runtime instance copy which can be nil if the runtime never got created in the first place. instead it now uses the reconciler copy which is always set because its based on the kubernetes object

Signed-off-by: Jakob Möller <contact@jakob-moeller.com>