Publish static kro manifests on release (#820)
* Publish static kro manifests on release

The industry tends to have different descriptions for
helm and plain k8s manifests with kustomize.

We can notice it in argoCD where they maintain the
[manifests](https://github.com/argoproj/argo-cd/blob/master/manifests/core-install.yaml)
as well as the helm [chart](https://github.com/argoproj/argo-helm/tree/main/charts/argo-cd)
(even in a different repository).

When we first tried to generate static manifests using this technique,
during the review, a concern was raised about the additional cognitive
load and maintenance burden having 2 different deployment methods would introduce.

This commit aims at providing static bundles to benefit from static manifests without
the additional maintenance burden.
It keeps Helm as the source of truth for deployment description and picks a list of well-known
values combination to generate static manifests out of it.

* Add make commands to install static manifests in kind

* Install static manifests in kro-system

chainsaw tests are using kro-system that is a standard
practice for operator namespaces.
Follow this naming convention

* Add make commands to run plain manifests e2e tests

* Add cluster-roles for chainsaw tests

chainsaw tests are run with unrestricted kro.
to tests raw manifests generation, we want to ensure that
it works for least-privilege mode.

in the least-privilege mode (core-install) kro does not have permission
to manage individual resources and hence e2e tests fails.

This commits extends Kro's privileges in end to end tests to
ensure they pass. It reflects how users should use KRO when using
the `rbac.mode=aggregation` in helm

* Update installation docs

* Address review comments

* Run 1 raw manifest in CI

Run the least privileged install in CI to ensure raw manifests works as expected

* Fix remaining failing chainsaw cleanup

* Move raw manifests e2e tests to dedicated script